Unbreak build.

+++ bench/BenchString.hs
@@ -3,6 +3,7 @@
 module Main where
 
 import qualified Criterion.Main as C
+import GHC.Int
 import Control.DeepSeq
 import Data.List (foldl')
 import qualified Y.String as S
@@ -30,16 +31,16 @@ benchLines text name
 benchDrop :: S.YiString -> String -> C.Benchmark
 benchDrop text name
     = C.bench name
-    $ C.nf (\x -> foldr S.drop x (replicate 1000 (1 :: Int))) text
+    $ C.nf (\x -> foldr S.drop x (replicate 1000 (1 :: Int64))) text
 
 benchTake :: S.YiString -> String -> C.Benchmark
 benchTake text name
     = C.bench name
-    $ C.nf (\x -> foldr S.take x [1000, 999 .. 1 :: Int]) text
+    $ C.nf (\x -> foldr S.take x [1000, 999 .. 1 :: Int64]) text
 
 benchSplitAt text name
     = C.bench name
-    $ C.nf (\x -> foldr ((fst .) . S.splitAt) x [1000, 999 .. 1 :: Int]) text
+    $ C.nf (\x -> foldr ((fst .) . S.splitAt) x [1000, 999 .. 1 :: Int64]) text
 
 main :: IO ()
 main = C.defaultMain
